Moved finddata into setup.py itself to fix running it in tox.
